unit tests and additional ignores
