+++ .gitignore
@@ -7,3 +7,5 @@ dist/*
 dist
 *.egg-info
 build
+.eggs
+
