+++ setup.py
@@ -8,6 +8,13 @@ here = path.abspath(path.dirname(__file__))
 with open(path.join(here, 'README.rst'), encoding='utf-8') as f:
     long_description = f.read()
 
+
+import unittest
+def my_test_suite():
+    test_loader = unittest.TestLoader()
+    test_suite = test_loader.discover('nerd/tests', pattern='test_*.py')
+    return test_suite
+
 setup(
     name='entity-fishing-client',
     version='0.4.0',
@@ -29,5 +36,7 @@ setup(
     install_requires=['requests', 'zenlog'],
     packages=['nerd'],
     # package_dir={'nerd': 'entity-fishing_client'},
-    zip_safe=False
+    zip_safe=False,
+    test_suite='setup.my_test_suite'
 )
+
